BSP2 support, based on an initial patch by Spoike.
Closes: https://sourceforge.net/p/uhexen2/feature-requests/36/

+++ engine/h2shared/progs.h
@@ -43,7 +43,7 @@ typedef struct edict_s
 	link_t		area;			/* linked to a division node or leaf */
 
 	int		num_leafs;
-	short		leafnums[MAX_ENT_LEAFS];
+	int		leafnums[MAX_ENT_LEAFS];
 
 	entity_state_t	baseline;
